drive_test: added drive control testing code

+++ drive_controls/src/lib.rs
@@ -11,16 +11,16 @@ pub fn arcade(x: i32, y: i32) -> (i32, i32) {
 }
 
 /// Daniels magic number for nice, smooth and exponential controls (`12000 = 1024a^{1}`)
-const DMN: f64 = 12.71875;
+const DMN: f32 = 12.71875;
 
 /// Passes x (`-1..=1`) through daniel's algorithm to produce an exponential voltage from `-12000..=12000`
-pub fn exp_daniel(x: f64) -> f64 {
+pub fn exp_daniel(x: f32) -> f32 {
     (1024.0 * maths::powf(DMN, maths::absf(x)) - 1024.0) // main part of the equation
         * maths::signumf(x) // to maintain the sign
 }
 
 /// Course corrects the x and y values (`-12000..=12000`) based on the interial sensor yaw
-pub fn course_correct(x: f64, y: f64, yaw: f64) -> (f64, f64) {
+pub fn course_correct(x: f32, y: f32, yaw: f32) -> (f32, f32) {
     // find the angle that the x and y make through the origin
     let angle = maths::atan(x / (y + 1.0));
